Add builder opts
